requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
wtforms-validators==1.0.0
Bcrypt-Flask==1.0.2
Flask-Bcrypt==1.0.1
bcrypt==4.1.1
azure-cosmos==4.5.1
requests==2.31.0

